+++ src/parser.ts
@@ -6,9 +6,9 @@ export function parseCommand(options: {
     rootCommandName: string,
     initOptions: YaclilOptions,
     argv: string[],
-}): ParsedCommand {
+}) {
     const parser = new Parser(options);
     parser.parse();
     const parsed = parser.parsedObject;
-    return parsed;
+    return { parsed, helpCalled: parser.helpCalled };
 }